Ditch old withdrawal
