+++ src/Hedwig/ClientApp/src/containers/Enrollment/Detail/EnrollmentDetail.tsx
@@ -79,7 +79,7 @@ export default function EnrollmentDetail({
 					</div>
 					<Button
 						text="Withdraw"
-						href={`/roster/sites/${siteId}/enrollments/${enrollment.id}/withdraw-new`}
+						href={`/roster/sites/${siteId}/enrollments/${enrollment.id}/withdraw`}
 						className="margin-right-0"
 					/>
 				</div>
